[linux] [update] urt_time is now long long
int64_t is long on 64 bit machines while its long long on 32 bit ones.
This caused confusion for printf.

+++ sys/linux/include/urt_sys_time.h
@@ -27,7 +27,7 @@
 
 URT_DECL_BEGIN
 
-typedef int64_t urt_time;
+typedef long long urt_time;
 
 static inline urt_time urt_get_time(void)
 {
